modularize "time_ago" logic
