+++ src/components/review/Review.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
     import { _ } from "svelte-i18n";
+    import { time_ago } from "../../utils/time_ago";
     
     export let name: string;
     export let body: string;
@@ -8,31 +9,8 @@
     export let short_body_char_limit: number = 150;
     export let link: string = "/review";
 
-    let time_ago = (created_at: string): string => {
-        let date = new Date(created_at);
-        let todayDate = new Date();
-        let timeDiff = Math.abs(todayDate.getTime() - date.getTime());
-        let diffDays = Math.floor(timeDiff / (1000 * 3600 * 24));
-        
-        if (diffDays === 0) {
-            return $_("today");
-        } else if (diffDays > 0 && diffDays < 7) {
-            // use diffDays as a key to get the correct translation
-            let message = diffDays === 1 ? "day_ago" : "days_ago";
-             return `${$_(message, { values: { days: diffDays } })}`;
-        } else if (diffDays >= 7 && diffDays < 30) {
-            let message = Math.floor(diffDays / 7) === 1 ? "week_ago" : "weeks_ago";
-            return `${$_(message, { values: { weeks: Math.floor(diffDays / 7) } })}`;
-        } else if (diffDays >= 30 && diffDays < 365) {
-            let message = Math.floor(diffDays / 30) === 1 ? "month_ago" : "months_ago";
-            return `${$_(message, { values: { months: Math.floor(diffDays / 30) } })}`;
-        } else if (diffDays >= 365) {
-            let message = Math.floor(diffDays / 365) === 1 ? "year_ago" : "years_ago";
-            return `${$_(message, { values: { years: Math.floor(diffDays / 365) } })}`;
-        }
-    };
-
-    let shortBody: string = body.slice(0, short_body_char_limit).trim();
+    let shortBody: string = body.slice(0, short_body_char_limit).trim();  
+    let {message, values} = time_ago(created_at);
   </script>
   
   <section class="review-container" data-testid="container">
@@ -52,7 +30,7 @@
             {/if}
         </div>
     </div>
-    <div class="review-date" data-testid="date">{time_ago(created_at)}</div>
+    <div class="review-date" data-testid="date">{$_(message, { values })}</div>
   </section>
   
   <style>
